requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.0
charset-normalizer==3.4.1
cryptography==43.0.0
defusedxml==0.7.1
Django==5.1.6
django-background-tasks==1.2.8
django-filter==24.3
django-oauth-toolkit==2.4.0
django-rest-framework-social-oauth2==1.2.0
django-storages==1.14.4
djangorestframework==3.15.2
google==3.0.0
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-core==2.4.1
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-oauth==1.0.1
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
idna==3.10
jwcrypto==1.5.6
Markdown==3.7
oauthlib==3.2.2
pillow==10.4.0
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.2.1
python3-openid==3.2.0
pytube==15.0.0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.17.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
yt-dlp==2025.1.15
